Btree: Add printer and clang-format
The btree printer is supposed to output a YML file describing the
current tree struct. It is supposed to work in combination with a tool
placed in `tools/` which transforms a given yml file to dot which could
be rendered using `dot`. Works fine.

![Supposed to be an image](https://i.ibb.co/Jr0TnSS/tree.png)

PS:

I know about the length of the diff that formed and I am sorry :).

Signed-off-by: Kristiyan Stoimenov <kristoimenov@gmail.com>