Move unit tests to corresponding directories
Move all unit tests out of the `tests/` dir and put them inside the
directory they belong.

e.g `TestBtree.cpp` → `storage/btree/Btree.cpp`

Nevertheless, they still compile and execute as before in the form of
unit tests.

Slight modification was made for the configuration of turning on unit
tests - `export EUGENE_BUILD_TESTS=1` is required in order for `eu test`
to recognise this option and behave as expected.

Signed-off-by: Kristiyan Stoimenov <kristoimenov@gmail.com>