Btree: Separate Config from Btree header
Config is moved to a separate file and changed from a concept to a
regular POD struct. This is done due to the fact that this enables us to
keep the default config at the exact same place as the POD definition,
whereas the struct required us to define a separate class which followed the required signature. There is still a concept which is defined to just check the parent of the passed Config class type.

In order to improve clarity a header which defines common suffixes was
added as well. It is used in the Config header.

Signed-off-by: Kristiyan Stoimenov <kristoimenov@gmail.com>