Btree: Make btree persistent
Add additional structure which helps with managing the btree on disk -
`Header`. When constructing a Btree with given `Config` and file
locations for the contents and header, try to load the header inside the
provided location if one is present and create a "fresh" tree only if
that is not the case.

Additional unit test added in the `Btree.cpp` - `Persistent tree`, which
validates the correctness of the current implementation.

Signed-off-by: Kristiyan Stoimenov <kristoimenov@gmail.com>