Btree: Fix segfault for some data types
Due to the fact that libnop appends additional bytes as
separators during serialization, not the whole 4K bytes
of the page were accessible to the btree nodes. Therefore,
the NUM_RECORDS_* constants were incorrect and were causing
undefined behaviour.

This patch provides a quick-and-dirty fix with binary search
at runtime for the most appropriate values. A future improvement
would be to at least make them compile-time operations.

Signed-off-by: Kristiyan Stoimenov <kristoimenov@gmail.com>